Minimum functioning version.
Add all base code to enable connection with
proxy srv. At this point the cli is able to
connect to a srv using net_id. The cli listens
to a static subnet and connects immediatley
Upon receving a matching beacon. This
version functions more or less perfectly
when using ext ADV, and partially with
legacy ADV. The latter case struggles to
make the connection while other adv- events
happens while the connetion is triggered.

Signed-off-by: Anders Storrø <anders.storro@nordicsemi.no>