Added base RX parsing and TX capabillities.
This commit adds the parsing of incoming
notification in order to sort them to their
respective types. It also adds the basis to
enable sending proxy cfg messages to the
server. At this point the client is connected to
the other layers whit regard to rx messages.
This is important to note since this is the
point where we expose the client to the rest
of the stack.

Signed-off-by: Anders Storrø <anders.storro@nordicsemi.no>